+++ samples/Avalonia/Demo.FluentContentDialog/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using ReactiveUI;
+
+namespace Demo.FluentContentDialog;
+
+public class ViewModelBase : ReactiveObject
+{
+}