+++ samples/Avalonia/Demo.FluentMessageBoxTaskDialog/App.axaml.cs
@@ -7,7 +7,7 @@ using HanumanInstitute.MvvmDialogs.Avalonia;
 using Microsoft.Extensions.Logging;
 using Splat;
 
-namespace Demo.MessageBox;
+namespace Demo.FluentContentDialog;
 
 public class App : Application
 {
@@ -22,7 +22,7 @@ public class App : Application
             new DialogManager(
                 viewLocator: new ViewLocator(),
                 logger: loggerFactory.CreateLogger<DialogManager>(),
-                dialogFactory: new DialogFactory().AddMessageBox()),
+                dialogFactory: new DialogFactory().AddFluent(FluentMessageBoxType.TaskDialog)),
             viewModelFactory: x => Locator.Current.GetService(x)));
 
         SplatRegistrations.Register<MainWindowViewModel>();