+++ samples/Avalonia/Demo.FluentTaskDialog/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using ReactiveUI;
+
+namespace Demo.FluentTaskDialog;
+
+public class ViewModelBase : ReactiveObject
+{
+}