+++ src/MvvmDialogs.Avalonia.MessageBox/GlobalUsings.cs
@@ -1,0 +1,6 @@
+global using System;
+global using System.Threading.Tasks;
+global using Avalonia;
+global using Avalonia.Controls;
+global using MessageBox.Avalonia;
+global using MessageBox.Avalonia.Enums;