+++ src/MvvmDialogs.Avalonia.MessageBox/IMessageBoxApi.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-using MessageBox.Avalonia.Enums;
 
 namespace HanumanInstitute.MvvmDialogs.Avalonia.MessageBox;