+++ src/MvvmDialogs.Avalonia.MessageBox/MessageBoxApiSettings.cs
@@ -1,6 +1,4 @@
-﻿using Avalonia.Controls;
-using MessageBox.Avalonia.Enums;
-
+﻿
 namespace HanumanInstitute.MvvmDialogs.Avalonia.MessageBox;
 
 internal class MessageBoxApiSettings